♻️ refactor: 수정하기 버튼 reply 상태 아닐때 보이도록
