+++ src/components/ui/atoms/question-box/AnswerBox.jsx
@@ -4,9 +4,12 @@ import styled from 'styled-components';
 
 import PROFILE_SAMPLE from '@components/ui/atoms/profile-sample';
 
+import { useToggle } from '@hooks/useToggle';
 import { feedCardType } from '@utils/card-type/feedCardType';
 import { timeStamp } from '@utils/time/timeStamp';
 
+import EditButton from '../Button/edit-button/EditButton';
+import Reaction from '../Reaction/Reaction';
 import ReplyTypeSwitch from '../ReplyTypeSwitch.jsx/ReplyTypeSwitch';
 
 // read, edit, reply type에 맞게 component 불러온다.
@@ -15,23 +18,39 @@ const AnswerBox = ({ answer, userProfile = PROFILE_SAMPLE, userName }) => {
   const [editTypeState, setEditTypeState] = useState({
     isEdit: false,
   });
+  const [isEdit, setIsEdit] = useToggle();
+
+  const handleEditToggle = () => {
+    setIsEdit();
+  };
 
   return (
-    <StAnswerBox>
-      <img src={userProfile} alt='프로필' />
-      <StAnswer>
-        <StUser>
-          <span className='name'>{userName}</span>
-          {answer !== null ? <span className='time'>{timeStamp(answer?.createdAt)}</span> : null}
-        </StUser>
-        <ReplyTypeSwitch
-          type={feedCardType(answer)}
-          value={answer?.content}
-          editTypeState={editTypeState}
-          isRejected={answer?.isRejected}
-        />
-      </StAnswer>
-    </StAnswerBox>
+    <>
+      <StAnswerBox>
+        <img src={userProfile} alt='프로필' />
+        <StAnswer>
+          <StUser>
+            <span className='name'>{userName}</span>
+            {answer !== null ? <span className='time'>{timeStamp(answer?.createdAt)}</span> : null}
+          </StUser>
+          <ReplyTypeSwitch
+            type={feedCardType(answer)}
+            value={answer?.content}
+            editTypeState={editTypeState}
+            isRejected={answer?.isRejected}
+          />
+        </StAnswer>
+      </StAnswerBox>
+      <StBottom>
+        <StLine />
+        {feedCardType(answer) !== 'reply' ? (
+          <StReactionAndEdit>
+            <Reaction likeCount={answer?.like} />
+            <EditButton onClickEdit={handleEditToggle} isEdit={isEdit} />
+          </StReactionAndEdit>
+        ) : null}
+      </StBottom>
+    </>
   );
 };
 
@@ -88,4 +107,25 @@ const StUser = styled.div`
     font-weight: 500;
     line-height: 18px; /* 128.571% */
   }
+`;
+
+const StBottom = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 16px;
+  align-self: stretch;
+`;
+
+const StReactionAndEdit = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  align-self: stretch;
+`;
+
+const StLine = styled.div`
+  height: 1px;
+  align-self: stretch;
+  background: ${({ theme }) => theme.color.Grayscale[30]};
 `;
