+++ src/components/ui/molecules/feed-card/FeedCard.jsx
@@ -1,36 +1,18 @@
 import styled from 'styled-components';
 
 import Badge from '@components/ui/atoms/Badge/Badge';
-import EditButton from '@components/ui/atoms/Button/edit-button/EditButton';
 import AnswerBox from '@components/ui/atoms/question-box/AnswerBox';
 import QuestionBox from '@components/ui/atoms/question-box/QuestionBox';
-import Reaction from '@components/ui/atoms/Reaction/Reaction';
 
-import { useToggle } from '@hooks/useToggle';
 import { feedCardType } from '@utils/card-type/feedCardType';
 import { timeStamp } from '@utils/time/timeStamp';
 
 const FeedCard = ({ answerResult, userName, userProfile }) => {
-  const [isEdit, setIsEdit] = useToggle();
-
-  const handleEditToggle = () => {
-    setIsEdit();
-  };
-
   return (
     <StFeedCard>
       <Badge value={feedCardType(answerResult?.answer)} />
       <QuestionBox question={answerResult?.content} elapsedTime={timeStamp(answerResult?.createdAt)} />
       <AnswerBox answer={answerResult?.answer} userProfile={userProfile} userName={userName} />
-      <StBottom>
-        <StLine />
-        <StReactionAndEdit>
-          <Reaction likeCount={answerResult?.like} />
-          {feedCardType(answerResult?.answer) === 'edit' ? (
-            <EditButton onClickEdit={handleEditToggle} isEdit={isEdit} />
-          ) : null}
-        </StReactionAndEdit>
-      </StBottom>
     </StFeedCard>
   );
 };
@@ -51,25 +33,4 @@ const StFeedCard = styled.div`
   background: ${({ theme }) => theme.color.Grayscale[10]};
 
   box-shadow: ${({ theme }) => theme.shadow[10]};
-`;
-
-const StBottom = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 16px;
-  align-self: stretch;
-`;
-
-const StReactionAndEdit = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  align-self: stretch;
-`;
-
-const StLine = styled.div`
-  height: 1px;
-  align-self: stretch;
-  background: ${({ theme }) => theme.color.Grayscale[30]};
 `;
